fix : 매칭 시(출석한 경우) 파트너 반영 되도록 추가
